refactor(users): updated variable names for filename and function
Updated the variable names for filename and function to match the naming convention. This makes the code more consistent and easier to read. Also added validation for token in login response.
